[@ember/service] add support for ember-octane @inject decorator
