+++ types/ember__service/index.d.ts
@@ -12,10 +12,9 @@ export default class Service extends EmberObject {}
  * Creates a property that lazily looks up a service in the container. There
  * are no restrictions as to what objects a service can be injected into.
  */
-export function inject(): ComputedProperty<Service>;
-export function inject<K extends keyof Registry>(
-    name: K
-): ComputedProperty<Registry[K]>;
+export function inject(): PropertyDecorator & ComputedProperty<Service>; // @inject() foo, foo: inject()
+export function inject(target: Object, propertyKey: string | symbol): void; // @inject foo
+export function inject<K extends keyof Registry>(name: K): PropertyDecorator & ComputedProperty<Registry[K]>; // @inject('store') foo      @inject() foo
 
 // A type registry for Ember `Service`s. Meant to be declaration-merged so
 // string lookups resolve to the correct type.
